Feature Update
Backspace:  Press 'B' while giving input to have a backspace.

I will try to add options for moving the cursor (and having a backspace or add new characters in the middle or at the start of the input).I don't know if I will be able to do that, but I will try.